restructured layout and configured auth routes
