+++ app/(auth)/layout.tsx
@@ -1,0 +1,7 @@
+export default async function AuthLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <>{children}</>;
+}
