+++ app/(customer-only)/_components/account/tabs/phone/sections/PhoneSection.tsx
@@ -21,7 +21,7 @@ export default function PhoneSection() {
         <div className="flex items-center space-x-4">
           <p className="text-sm text-gray-500">{phone as string}</p>
           <Button variant="ghost" onClick={() => setIsEditing(true)}>
-            Edit
+            {phone ? "Edit" : "Add"}
           </Button>
         </div>
       ) : isEditing ? (
