+++ app/(customer-only)/customer/page.tsx
@@ -1,0 +1,7 @@
+export default function CustomerHome() {
+  return (
+    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
+      Customer Home
+    </div>
+  );
+}
